requirements.txt
langchain==0.3.27
langchain-openai==0.3.32
qdrant-client==1.15.1
openai==1.109.1
fastapi==0.118.0
uvicorn==0.34.0
httpx==0.28.1
python-dotenv==1.2.2
pydantic==2.11.10